+++ canopen_analyzer/dox/documentation/html/classeds__parser_1_1eds__parser.js
@@ -1,11 +1,13 @@
 var classeds__parser_1_1eds__parser =
 [
     [ "__init__", "classeds__parser_1_1eds__parser.html#a50e50f7214b77463663d40b1d793543e", null ],
+    [ "build_entry_map", "classeds__parser_1_1eds__parser.html#a6e6c0bf72427b79f64f2a42af461f080", null ],
     [ "build_name_map", "classeds__parser_1_1eds__parser.html#a8013008a9a6ef94e9571b038c9eb1fec", null ],
     [ "log_pdo_mapping_consistency", "classeds__parser_1_1eds__parser.html#a55ddf37eb28c77078dafb3d7659fcee5", null ],
     [ "parse_pdo_map", "classeds__parser_1_1eds__parser.html#a302fbdb20e8a40ceb107b4ee0c9e2708", null ],
     [ "cob_name_overrides", "classeds__parser_1_1eds__parser.html#a428159b615b67863252e4b5463fa3771", null ],
     [ "eds_path", "classeds__parser_1_1eds__parser.html#ac38b8f2a825269765b2858c7f2155aec", null ],
+    [ "entry_map", "classeds__parser_1_1eds__parser.html#a2032165c96f39d9aa1502709f38acd3e", null ],
     [ "log", "classeds__parser_1_1eds__parser.html#ab09a63eb35b270b5cdbead1983ebdccb", null ],
     [ "name_map", "classeds__parser_1_1eds__parser.html#a379b3d2f78238184664fb872f44c6f96", null ],
     [ "pdo_map", "classeds__parser_1_1eds__parser.html#ac960737f47148af2dd1db69a47a9c803", null ],