+++ canopen_analyzer/dox/documentation/html/search/all_2.js
@@ -1,11 +1,13 @@
 var searchData=
 [
-  ['bitrate_69',['bitrate',['../classbus__stats_1_1bus__stats.html#a9ba628ef53aa2760ee08afab49765d42',1,'bus_stats::bus_stats']]],
-  ['build_5fname_5fmap_70',['build_name_map',['../classeds__parser_1_1eds__parser.html#a8013008a9a6ef94e9571b038c9eb1fec',1,'eds_parser::eds_parser']]],
-  ['bus_71',['bus',['../classcanopen__sniffer_1_1canopen__sniffer.html#a411bd59fecc82fdfda0e5c9d7da7447f',1,'canopen_sniffer::canopen_sniffer']]],
-  ['bus_5flabels_72',['bus_labels',['../classdisplay__gui_1_1CANopenMainWindow.html#adf0ff66d2d75c14e64b9f0738132fb27',1,'display_gui::CANopenMainWindow']]],
-  ['bus_5fstats_73',['bus_stats',['../namespacebus__stats.html',1,'bus_stats'],['../classbus__stats_1_1bus__stats.html',1,'bus_stats']]],
-  ['bus_5fstats_2epy_74',['bus_stats.py',['../bus__stats_8py.html',1,'']]],
-  ['bus_5fstats_5ftable_75',['bus_stats_table',['../classdisplay__tui_1_1display__tui.html#a13322e7aa50616f0af270c661d2ffb1c',1,'display_tui::display_tui']]],
-  ['bytes_5fto_5fhex_76',['bytes_to_hex',['../namespaceanalyzer__defs.html#aa969db480a6f6ee9e64524d308684239',1,'analyzer_defs']]]
+  ['bit_5flength_5fby_5ftype_73',['BIT_LENGTH_BY_TYPE',['../namespaceanalyzer__defs.html#af5bae2689cd20678b69a1d8018195341',1,'analyzer_defs']]],
+  ['bitrate_74',['bitrate',['../classbus__stats_1_1bus__stats.html#a9ba628ef53aa2760ee08afab49765d42',1,'bus_stats::bus_stats']]],
+  ['build_5fentry_5fmap_75',['build_entry_map',['../classeds__parser_1_1eds__parser.html#a6e6c0bf72427b79f64f2a42af461f080',1,'eds_parser::eds_parser']]],
+  ['build_5fname_5fmap_76',['build_name_map',['../classeds__parser_1_1eds__parser.html#a8013008a9a6ef94e9571b038c9eb1fec',1,'eds_parser::eds_parser']]],
+  ['bus_77',['bus',['../classcanopen__sniffer_1_1canopen__sniffer.html#a411bd59fecc82fdfda0e5c9d7da7447f',1,'canopen_sniffer::canopen_sniffer']]],
+  ['bus_5flabels_78',['bus_labels',['../classdisplay__gui_1_1CANopenMainWindow.html#adf0ff66d2d75c14e64b9f0738132fb27',1,'display_gui::CANopenMainWindow']]],
+  ['bus_5fstats_79',['bus_stats',['../namespacebus__stats.html',1,'bus_stats'],['../classbus__stats_1_1bus__stats.html',1,'bus_stats']]],
+  ['bus_5fstats_2epy_80',['bus_stats.py',['../bus__stats_8py.html',1,'']]],
+  ['bus_5fstats_5ftable_81',['bus_stats_table',['../classdisplay__tui_1_1display__tui.html#a13322e7aa50616f0af270c661d2ffb1c',1,'display_tui::display_tui']]],
+  ['bytes_5fto_5fhex_82',['bytes_to_hex',['../namespaceanalyzer__defs.html#aa969db480a6f6ee9e64524d308684239',1,'analyzer_defs']]]
 ];