+++ canopen_analyzer/dox/documentation/html/search/functions_10.js
@@ -1,9 +1,9 @@
 var searchData=
 [
-  ['save_5fprocessed_5fframe_406',['save_processed_frame',['../classprocess__frames_1_1process__frames.html#abab717783b0e365598b1f9c9205577d2',1,'process_frames::process_frames']]],
-  ['send_5fraw_5fpdo_407',['send_raw_pdo',['../classcanopen__sniffer_1_1canopen__sniffer.html#af5700ac79e73b5f715f582e443ccc969',1,'canopen_sniffer::canopen_sniffer']]],
-  ['send_5fsdo_5fdownload_408',['send_sdo_download',['../classcanopen__sniffer_1_1canopen__sniffer.html#aa8d11ec8573e1db81498322bba468a2d',1,'canopen_sniffer::canopen_sniffer']]],
-  ['send_5fsdo_5fupload_5frequest_409',['send_sdo_upload_request',['../classcanopen__sniffer_1_1canopen__sniffer.html#a097369338fe5940560a0d109f2416bc5',1,'canopen_sniffer::canopen_sniffer']]],
-  ['set_5fstart_5ftime_410',['set_start_time',['../classbus__stats_1_1bus__stats.html#a09ae5bd534d8dbec99a913ed75fe9848',1,'bus_stats::bus_stats']]],
-  ['stop_411',['stop',['../classbus__stats_1_1bus__stats.html#a26ca7c1c7fcdd35378e7be97727047a6',1,'bus_stats.bus_stats.stop()'],['../classcanopen__sniffer_1_1canopen__sniffer.html#a2342b933b01f118be030e93341908480',1,'canopen_sniffer.canopen_sniffer.stop()'],['../classdisplay__cli_1_1display__cli.html#a26ca7c1c7fcdd35378e7be97727047a6',1,'display_cli.display_cli.stop()'],['../classdisplay__gui_1_1GUIUpdateWorker.html#a26ca7c1c7fcdd35378e7be97727047a6',1,'display_gui.GUIUpdateWorker.stop()'],['../classprocess__frames_1_1process__frames.html#a26ca7c1c7fcdd35378e7be97727047a6',1,'process_frames.process_frames.stop()']]]
+  ['save_5fprocessed_5fframe_420',['save_processed_frame',['../classprocess__frames_1_1process__frames.html#abab717783b0e365598b1f9c9205577d2',1,'process_frames::process_frames']]],
+  ['send_5fraw_5fpdo_421',['send_raw_pdo',['../classcanopen__sniffer_1_1canopen__sniffer.html#af5700ac79e73b5f715f582e443ccc969',1,'canopen_sniffer::canopen_sniffer']]],
+  ['send_5fsdo_5fdownload_422',['send_sdo_download',['../classcanopen__sniffer_1_1canopen__sniffer.html#aa8d11ec8573e1db81498322bba468a2d',1,'canopen_sniffer::canopen_sniffer']]],
+  ['send_5fsdo_5fupload_5frequest_423',['send_sdo_upload_request',['../classcanopen__sniffer_1_1canopen__sniffer.html#a097369338fe5940560a0d109f2416bc5',1,'canopen_sniffer::canopen_sniffer']]],
+  ['set_5fstart_5ftime_424',['set_start_time',['../classbus__stats_1_1bus__stats.html#a09ae5bd534d8dbec99a913ed75fe9848',1,'bus_stats::bus_stats']]],
+  ['stop_425',['stop',['../classbus__stats_1_1bus__stats.html#a26ca7c1c7fcdd35378e7be97727047a6',1,'bus_stats.bus_stats.stop()'],['../classcanopen__sniffer_1_1canopen__sniffer.html#a2342b933b01f118be030e93341908480',1,'canopen_sniffer.canopen_sniffer.stop()'],['../classdisplay__cli_1_1display__cli.html#a26ca7c1c7fcdd35378e7be97727047a6',1,'display_cli.display_cli.stop()'],['../classdisplay__gui_1_1GUIUpdateWorker.html#a26ca7c1c7fcdd35378e7be97727047a6',1,'display_gui.GUIUpdateWorker.stop()'],['../classprocess__frames_1_1process__frames.html#a26ca7c1c7fcdd35378e7be97727047a6',1,'process_frames.process_frames.stop()']]]
 ];