+++ canopen_analyzer/dox/documentation/html/search/namespaces_1.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['bus_5fstats_309',['bus_stats',['../namespacebus__stats.html',1,'']]]
+  ['bus_5fstats_318',['bus_stats',['../namespacebus__stats.html',1,'']]]
 ];