+++ canopen_analyzer/dox/documentation/html/search/variables_1.js
@@ -1,8 +1,8 @@
 var searchData=
 [
-  ['app_431',['App',['../namespacedisplay__tui.html#a0ac8512426efbd5aed951550c0939686',1,'display_tui']]],
-  ['app_5fname_432',['APP_NAME',['../namespaceanalyzer__defs.html#a54f4259535eb0ca336b0b80a46aaf3bf',1,'analyzer_defs']]],
-  ['app_5forg_433',['APP_ORG',['../namespaceanalyzer__defs.html#acd590119f951363e7f34cfc58f80e6bb',1,'analyzer_defs']]],
-  ['axis_5fx_434',['axis_x',['../classdisplay__gui_1_1MultiRateLineWidget.html#ac58cbf59937d456275d30201980d6b10',1,'display_gui::MultiRateLineWidget']]],
-  ['axis_5fy_435',['axis_y',['../classdisplay__gui_1_1MultiRateLineWidget.html#a3bbe369a3f3690dff3c09d8aa742295a',1,'display_gui::MultiRateLineWidget']]]
+  ['app_446',['App',['../namespacedisplay__tui.html#a0ac8512426efbd5aed951550c0939686',1,'display_tui']]],
+  ['app_5fname_447',['APP_NAME',['../namespaceanalyzer__defs.html#a54f4259535eb0ca336b0b80a46aaf3bf',1,'analyzer_defs']]],
+  ['app_5forg_448',['APP_ORG',['../namespaceanalyzer__defs.html#acd590119f951363e7f34cfc58f80e6bb',1,'analyzer_defs']]],
+  ['axis_5fx_449',['axis_x',['../classdisplay__gui_1_1MultiRateLineWidget.html#ac58cbf59937d456275d30201980d6b10',1,'display_gui::MultiRateLineWidget']]],
+  ['axis_5fy_450',['axis_y',['../classdisplay__gui_1_1MultiRateLineWidget.html#a3bbe369a3f3690dff3c09d8aa742295a',1,'display_gui::MultiRateLineWidget']]]
 ];