+++ canopen_analyzer/dox/documentation/html/search/variables_2.js
@@ -1,7 +1,8 @@
 var searchData=
 [
-  ['bitrate_436',['bitrate',['../classbus__stats_1_1bus__stats.html#a9ba628ef53aa2760ee08afab49765d42',1,'bus_stats::bus_stats']]],
-  ['bus_437',['bus',['../classcanopen__sniffer_1_1canopen__sniffer.html#a411bd59fecc82fdfda0e5c9d7da7447f',1,'canopen_sniffer::canopen_sniffer']]],
-  ['bus_5flabels_438',['bus_labels',['../classdisplay__gui_1_1CANopenMainWindow.html#adf0ff66d2d75c14e64b9f0738132fb27',1,'display_gui::CANopenMainWindow']]],
-  ['bus_5fstats_5ftable_439',['bus_stats_table',['../classdisplay__tui_1_1display__tui.html#a13322e7aa50616f0af270c661d2ffb1c',1,'display_tui::display_tui']]]
+  ['bit_5flength_5fby_5ftype_451',['BIT_LENGTH_BY_TYPE',['../namespaceanalyzer__defs.html#af5bae2689cd20678b69a1d8018195341',1,'analyzer_defs']]],
+  ['bitrate_452',['bitrate',['../classbus__stats_1_1bus__stats.html#a9ba628ef53aa2760ee08afab49765d42',1,'bus_stats::bus_stats']]],
+  ['bus_453',['bus',['../classcanopen__sniffer_1_1canopen__sniffer.html#a411bd59fecc82fdfda0e5c9d7da7447f',1,'canopen_sniffer::canopen_sniffer']]],
+  ['bus_5flabels_454',['bus_labels',['../classdisplay__gui_1_1CANopenMainWindow.html#adf0ff66d2d75c14e64b9f0738132fb27',1,'display_gui::CANopenMainWindow']]],
+  ['bus_5fstats_5ftable_455',['bus_stats_table',['../classdisplay__tui_1_1display__tui.html#a13322e7aa50616f0af270c661d2ffb1c',1,'display_tui::display_tui']]]
 ];