+++ canopen_analyzer/dox/documentation/html/search/variables_a.js
@@ -1,9 +1,9 @@
 var searchData=
 [
-  ['max_5fcli_5fcmd_5fhistory_493',['MAX_CLI_CMD_HISTORY',['../namespaceanalyzer__defs.html#adfdb2aa7ec3d939c91e68a67f786ca9f',1,'analyzer_defs']]],
-  ['max_5fframes_494',['MAX_FRAMES',['../namespaceanalyzer__defs.html#a90270d61ef5cbcf52b1f4e63ee96cfbc',1,'analyzer_defs']]],
-  ['max_5fpoints_495',['max_points',['../classdisplay__gui_1_1MultiRateLineWidget.html#a043afb44cd3575db763c65fd445a5858',1,'display_gui::MultiRateLineWidget']]],
-  ['max_5fstats_5fshow_496',['MAX_STATS_SHOW',['../namespaceanalyzer__defs.html#aa6fda8207dcce3b56fd8ed0ff3ec827e',1,'analyzer_defs']]],
-  ['min_5fstats_5fshow_497',['MIN_STATS_SHOW',['../namespaceanalyzer__defs.html#a80c4d7f9c3a827951418fea4187ad0df',1,'analyzer_defs']]],
-  ['mode_5fcombo_498',['mode_combo',['../classdisplay__gui_1_1CANopenMainWindow.html#ae34ca3af9e134197585f22c51544b1ab',1,'display_gui::CANopenMainWindow']]]
+  ['max_5fcli_5fcmd_5fhistory_511',['MAX_CLI_CMD_HISTORY',['../namespaceanalyzer__defs.html#adfdb2aa7ec3d939c91e68a67f786ca9f',1,'analyzer_defs']]],
+  ['max_5fframes_512',['MAX_FRAMES',['../namespaceanalyzer__defs.html#a90270d61ef5cbcf52b1f4e63ee96cfbc',1,'analyzer_defs']]],
+  ['max_5fpoints_513',['max_points',['../classdisplay__gui_1_1MultiRateLineWidget.html#a043afb44cd3575db763c65fd445a5858',1,'display_gui::MultiRateLineWidget']]],
+  ['max_5fstats_5fshow_514',['MAX_STATS_SHOW',['../namespaceanalyzer__defs.html#aa6fda8207dcce3b56fd8ed0ff3ec827e',1,'analyzer_defs']]],
+  ['min_5fstats_5fshow_515',['MIN_STATS_SHOW',['../namespaceanalyzer__defs.html#a80c4d7f9c3a827951418fea4187ad0df',1,'analyzer_defs']]],
+  ['mode_5fcombo_516',['mode_combo',['../classdisplay__gui_1_1CANopenMainWindow.html#ae34ca3af9e134197585f22c51544b1ab',1,'display_gui::CANopenMainWindow']]]
 ];