+++ CHANGELOG.md
@@ -19,6 +19,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 - Updated frame simulator script to support DataTypes, segmented SDOs and maintaining local DB.
+- Added support for DataType in analyzer and handling of data type as per DataTypes.
+- Updated Bus utilization calculations.
+- Added TUI notifications for Send and Repeat.
 
 ## [v0.22.0] - 2026-01-03